+++ packages/terra-application-links/src/tabs/ApplicationTabs.jsx
@@ -172,7 +172,7 @@ class ApplicationTabs extends React.Component {
     });
 
     return (
-      <div {...customProps} className={cx(['application-tabs'])}>
+      <div {...customProps} className={cx(['tabs-wrapper'])}>
         <div
           className={cx(['tabs-container', { 'is-calculating': this.isCalculating }, alignment])}
           role="tablist"
@@ -182,6 +182,7 @@ class ApplicationTabs extends React.Component {
           <TabMenu location={location} isHidden={this.menuHidden}>
             {hiddenChildren}
           </TabMenu>
+          <div className={cx(['divider-after-last-tab'])} />
         </div>
       </div>
     );