+++ packages/terra-application-name/src/terra-dev-site/test/application-name/ApplicationMenuNameWrapping.test.jsx
@@ -5,8 +5,8 @@ import ApplicationMenuName from '../../../ApplicationMenuName';
 export default () => (
   <div style={{ width: '150px', backgroundColor: 'green' }}>
     <ApplicationMenuName
-      id="truncated"
-      title="TitleTitleTitleTitleTitleTitleTitle"
+      id="wrapping"
+      title="Title TitleTitleTitleTitleTitle Title Title Title Title Title"
       accessory={<Image alt="terra accessory" variant="rounded" src="https://github.com/cerner/terra-framework/raw/master/terra.png" height="26px" width="26px" />}
     />
   </div>