+++ packages/terra-hookshot/CHANGELOG.md
@@ -4,7 +4,7 @@ Changelog
 Unreleased
 ----------
 ### Removed
-* Vestigal fixed styling removed and default to absolute. 
+* Vestigial fixed styling removed and default to absolute. 
 
 5.0.0 - (January 25, 2019)
 ------------------