+++ packages/terra-hookshot/CHANGELOG.md
@@ -3,6 +3,9 @@ Changelog
 
 Unreleased
 ----------
+
+5.0.0 - (January 25, 2019)
+------------------
 ### Breaking Change
 * Updated terra-core dependencies